Extract HTTPIO
