+++ lib/cucumber/formatter/io.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 require 'net/http'
+require 'cucumber/formatter/http_io'
 
 module Cucumber
   module Formatter
@@ -35,40 +36,6 @@ module Cucumber
         raise "I can't write #{name} reports to a file - it has to be a directory" if File.file?(path)
         FileUtils.mkdir_p(path) unless File.directory?(path)
         File.absolute_path path
-      end
-      
-      class HTTPIO
-        def initialize(url, https_verify_mode=nil)
-          @url = url
-          @https_verify_mode = https_verify_mode
-          @closed = false
-          @body = ''
-        end
-        
-        def write(chunk)
-          @body += chunk
-        end
-        
-        def closed?
-          @closed
-        end
-        
-        def flush
-          uri = URI(@url)
-          req = Net::HTTP::Post.new(uri)
-          req.body = @body
-          http = Net::HTTP.new(uri.hostname, uri.port)
-          if uri.scheme == 'https'
-            http.use_ssl = true
-            http.verify_mode = @https_verify_mode if @https_verify_mode
-          end
-          res = http.request(req)
-          raise "Not OK" unless Net::HTTPOK === res
-        end
-        
-        def close
-          @closed = true
-        end
       end
     end
   end
